[#1567] moved classes
