+++ app/src/main/java/org/akvo/flow/presentation/form/view/groups/QuestionGroupFragmentFragment.kt
@@ -17,7 +17,7 @@
  * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.akvo.flow.presentation.form.view.ui.main
+package org.akvo.flow.presentation.form.view.groups
 
 import android.os.Bundle
 import android.view.LayoutInflater
