+++ app/src/main/res/layout/question_group_fragment.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.form.view.ui.main.QuestionGroupFragmentFragment">
+    tools:context=".presentation.form.view.ui.groups.QuestionGroupFragmentFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView2"
